change namespace and remove upsampling
